--- src/constants/api.ts
@@ -1,3 +1,0 @@
-export const OPENSEARCH_API_BASE_URL = "https://iss.ndl.go.jp/api/opensearch";
-
-export const THUMBNAIL_API_BASE_URL = "https://ndlsearch.ndl.go.jp/thumbnail";
